requirements.txt
amqp==5.3.1
anyascii==0.3.3
asgiref==3.9.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
defusedxml==0.7.1
Django==5.2.4
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-filter==25.1
django-formtools==2.5.1
django-htmx==1.23.2
django-modelcluster==6.4
django-permissionedforms==0.1
django-phonenumber-field==8.1.0
django-phonenumbers==1.0.1
django-recaptcha==4.1.0
django-six==1.0.5
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
excel-base==1.0.4
filetype==1.2.0
gunicorn==23.0.0
idna==3.10
isoweek==1.3.3
kombu==5.5.4
laces==0.1.2
openpyxl==3.1.5
packaging==25.0
phonenumbers==9.0.9
pillow==11.3.0
pillow_heif==1.0.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
screen==1.0.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
TimeConvert==3.0.13
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wagtail==7.0.1
wagtail-modeladmin==2.2.0
wcwidth==0.2.13
Willow==1.11.0
xlwt==1.3.0
